Split sensors to multiple files
