+++ custom_components/huckleberry/features/__init__.py
@@ -1,0 +1,1 @@
+"""Feature-oriented entity modules for Huckleberry."""
