+++ custom_components/huckleberry/switch.py
@@ -1,7 +1,6 @@
 """Switch platform for Huckleberry integration."""
 from __future__ import annotations
 
-import logging
 from typing import cast
 
 from homeassistant.components.switch import SwitchEntity
@@ -9,15 +8,10 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from huckleberry_api import HuckleberryAPI
-from huckleberry_api.firebase_types import FeedSide
-
-from . import HuckleberryDataUpdateCoordinator, HuckleberryEntryData
+from . import HuckleberryEntryData
 from .const import DOMAIN
-from .entity import HuckleberryBaseEntity
-from .models import HuckleberryChildProfile
-
-_LOGGER = logging.getLogger(__name__)
+from .features.nursing import build_nursing_switches
+from .features.sleep import build_sleep_switches
 
 
 async def async_setup_entry(
@@ -29,81 +23,8 @@ async def async_setup_entry(
     entry_data = cast(HuckleberryEntryData, hass.data[DOMAIN][entry.entry_id])
     entities: list[SwitchEntity] = []
 
-    for child in entry_data["children"]:
-        entities.append(HuckleberrySleepSwitch(entry_data["coordinator"], entry_data["api"], child))
-        entities.append(HuckleberryNursingSwitch(entry_data["coordinator"], entry_data["api"], child, "left"))
-        entities.append(HuckleberryNursingSwitch(entry_data["coordinator"], entry_data["api"], child, "right"))
+    entities.extend(build_sleep_switches(entry_data["coordinator"], entry_data["api"], entry_data["children"]))
+    entities.extend(build_nursing_switches(entry_data["coordinator"], entry_data["api"], entry_data["children"]))
 
     async_add_entities(entities)
 
-
-class HuckleberrySleepSwitch(HuckleberryBaseEntity, SwitchEntity):
-    """Switch to start or stop sleep tracking."""
-
-    def __init__(
-        self,
-        coordinator: HuckleberryDataUpdateCoordinator,
-        api: HuckleberryAPI,
-        child: HuckleberryChildProfile,
-    ) -> None:
-        super().__init__(coordinator, child)
-        self._api = api
-        self._attr_name = "Sleep tracking"
-        self._attr_unique_id = f"{self.child_uid}_sleep_tracking"
-        self._attr_icon = "mdi:sleep"
-
-    @property
-    def is_on(self) -> bool:
-        """Return true if sleep tracking is active and not paused."""
-        sleep_status = self.coordinator.get_sleep_status(self.child_uid)
-        timer = sleep_status.timer if sleep_status is not None else None
-        return bool(timer is not None and timer.active and not timer.paused)
-
-    async def async_turn_on(self, **kwargs: object) -> None:
-        """Start sleep tracking."""
-        _LOGGER.info("Starting sleep tracking for %s", self.child_name)
-        await self._api.start_sleep(self.child_uid)
-
-    async def async_turn_off(self, **kwargs: object) -> None:
-        """Complete sleep tracking."""
-        _LOGGER.info("Stopping sleep tracking for %s", self.child_name)
-        await self._api.complete_sleep(self.child_uid)
-
-
-class HuckleberryNursingSwitch(HuckleberryBaseEntity, SwitchEntity):
-    """Switch to start or stop nursing tracking for a specific side."""
-
-    def __init__(
-        self,
-        coordinator: HuckleberryDataUpdateCoordinator,
-        api: HuckleberryAPI,
-        child: HuckleberryChildProfile,
-        side: FeedSide,
-    ) -> None:
-        super().__init__(coordinator, child)
-        self._api = api
-        self._side: FeedSide = side
-        self._attr_name = f"Nursing {side}"
-        self._attr_unique_id = f"{self.child_uid}_nursing_{side}"
-        self._attr_icon = "mdi:baby-bottle" if side == "left" else "mdi:baby-bottle-outline"
-
-    @property
-    def is_on(self) -> bool:
-        """Return true if nursing tracking is active on this side."""
-        feed_status = self.coordinator.get_feed_status(self.child_uid)
-        timer = feed_status.timer if feed_status is not None else None
-        if timer is None or not timer.active or timer.paused:
-            return False
-
-        active_side = timer.activeSide or timer.lastSide
-        return active_side == self._side
-
-    async def async_turn_on(self, **kwargs: object) -> None:
-        """Start nursing tracking on this side."""
-        _LOGGER.info("Starting %s nursing for %s", self._side, self.child_name)
-        await self._api.start_nursing(self.child_uid, self._side)
-
-    async def async_turn_off(self, **kwargs: object) -> None:
-        """Complete nursing tracking and save to history."""
-        _LOGGER.info("Completing nursing for %s", self.child_name)
-        await self._api.complete_nursing(self.child_uid)
